rm bottomNav and add Header
